// web/src/api-access/generated/models/leaveCalendarEventResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Unified.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface LeaveCalendarEventResponse {
  id: string;
  leaveId?: number;
  eventId?: number;
  userId: string;
  type: string;
  sourceModule: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  color?: string | null;
  start: string;
  /** @nullable */
  end?: string | null;
  /** @nullable */
  timeZoneId?: string | null;
  allDay?: boolean;
  eventTypeCode: string;
  statusTypeCode: string;
  /** @nullable */
  cancelledAt?: string | null;
  /** @nullable */
  cancelledByUserId?: string | null;
  /** @nullable */
  cancellationReason?: string | null;
  leaveTypeId?: number;
  leaveTypeCode: string;
  isPaid?: boolean;
}